refactored a bit lara will finish tonight
