+++ README.md
@@ -54,3 +54,11 @@ Note: Use vscode to program. Android is just to launch the app on a mobile andro
 2. ionic serve
    - ionic will indicate the local host address
    - should be localhost:8100
+
+### How to run the app on an android phone
+
+1. npm i -g native-run (if first time building)
+2. `ionic capacitor add android`
+3. `ionic build`
+4. `npx cap add android`
+5. `npx cap open android`
