+++ ionic.config.json
@@ -1,7 +1,8 @@
 {
   "name": "pikaroute",
   "integrations": {
-    "cordova": {}
+    "cordova": {},
+    "capacitor": {}
   },
   "type": "angular"
 }
