+++ src/app/app.module.ts
@@ -12,6 +12,7 @@ import { AppRoutingModule } from './app-routing.module';
 import { Geolocation } from '@ionic-native/geolocation/ngx';
 import { NativeGeocoder } from '@ionic-native/native-geocoder/ngx';
 
+
 @NgModule({
   declarations: [AppComponent],
   entryComponents: [],
