--- src/app/google-api.service.ts
@@ -1,76 +1,0 @@
-import { Injectable } from '@angular/core';
-import { ElementRef } from '@angular/core';
-import { Geolocation } from '@ionic-native/geolocation/ngx';
-import { NativeGeocoder, NativeGeocoderResult, NativeGeocoderOptions } from '@ionic-native/native-geocoder/ngx';
-
-declare var google;
-
-@Injectable({
-  providedIn: 'root'
-})
-export class GoogleApiService {
-  constructor(
-    private geolocation: Geolocation,
-    private nativeGeocoder: NativeGeocoder
-  ) {
-  }
-
-  map: any;
-  address:string;
-
-  loadMap(mapElement: ElementRef, starting_latitude: number = 0.0, starting_longitude: number = 0.0) {
-    this.geolocation.getCurrentPosition().then((resp) => {
-        let latLng: any;
-      if(starting_latitude==0.0 && starting_longitude==0.0){
-        latLng = new google.maps.LatLng(resp.coords.latitude, resp.coords.longitude);
-      }else{
-        latLng = new google.maps.LatLng(starting_latitude,starting_longitude);
-      }
-      let mapOptions = {
-        center: latLng,
-        zoom: 15,
-        mapTypeId: google.maps.MapTypeId.ROADMAP
-      }
- 
-      this.getAddressFromCoords(resp.coords.latitude, resp.coords.longitude);
- 
-      this.map = new google.maps.Map(mapElement.nativeElement, mapOptions);
- 
-      this.map.addListener('tilesloaded', () => {
-        console.log('accuracy',this.map);
-        this.getAddressFromCoords(this.map.center.lat(), this.map.center.lng())
-      });
- 
-    }).catch((error) => {
-      console.log('Error getting location', error);
-    });
-  }
- 
-  getAddressFromCoords(lattitude, longitude) {
-    console.log("getAddressFromCoords "+lattitude+" "+longitude);
-    let options: NativeGeocoderOptions = {
-      useLocale: true,
-      maxResults: 5
-    };
- 
-    this.nativeGeocoder.reverseGeocode(lattitude, longitude, options)
-      .then((result: NativeGeocoderResult[]) => {
-        this.address = "";
-        let responseAddress = [];
-        for (let [key, value] of Object.entries(result[0])) {
-          if(value.length>0)
-          responseAddress.push(value);
- 
-        }
-        responseAddress.reverse();
-        for (let value of responseAddress) {
-          this.address += value+", ";
-        }
-        this.address = this.address.slice(0, -2);
-      })
-      .catch((error: any) =>{ 
-        this.address = "Address Not Available!";
-      });
- 
-  }
-}
