+++ src/app/google-apis/location.service.ts
@@ -1,0 +1,41 @@
+import { Injectable } from '@angular/core';
+import { NativeGeocoderOptions, NativeGeocoderResult, NativeGeocoder } from '@ionic-native/native-geocoder/ngx';
+
+@Injectable({
+  providedIn: 'root'
+})
+export class LocationService {
+
+  constructor(private nativeGeocoder: NativeGeocoder
+  ) { }
+
+  address: string;
+
+  getAddressFromCoords(latitude, longitude) {
+    console.log("getAddressFromCoords " + latitude + " " + longitude);
+    let options: NativeGeocoderOptions = {
+      useLocale: true,
+      maxResults: 5
+    };
+
+    this.nativeGeocoder.reverseGeocode(latitude, longitude, options)
+      .then((result: NativeGeocoderResult[]) => {
+        this.address = "";
+        let responseAddress = [];
+        for (let [key, value] of Object.entries(result[0])) {
+          if (value.length > 0)
+            responseAddress.push(value);
+
+        }
+        responseAddress.reverse();
+        for (let value of responseAddress) {
+          this.address += value + ", ";
+        }
+        this.address = this.address.slice(0, -2);
+      })
+      .catch((error: any) => {
+        this.address = "Address Not Available!";
+      });
+
+  }
+}
