+++ src/app/google-apis/map.service.spec.ts
@@ -1,0 +1,12 @@
+import { TestBed } from '@angular/core/testing';
+
+import { MapService } from './map.service';
+
+describe('MapService', () => {
+  beforeEach(() => TestBed.configureTestingModule({}));
+
+  it('should be created', () => {
+    const service: MapService = TestBed.get(MapService);
+    expect(service).toBeTruthy();
+  });
+});
